+++ lib/ethercat/driver/el2809.ex
@@ -2,6 +2,7 @@ defmodule EtherCAT.Driver.EL2809 do
   @moduledoc "Beckhoff EL2809 16-channel digital output, 24 V DC."
 
   @behaviour EtherCAT.Driver
+  alias EtherCAT.Endpoint
 
   @vendor_id 0x0000_0002
   @product_code 0x0AF9_3052
@@ -49,7 +50,14 @@ defmodule EtherCAT.Driver.EL2809 do
   end
 
   @impl true
-  def describe(_config), do: %{device_type: :digital_output, capabilities: [:set_output]}
+  def describe(_config) do
+    %{
+      device_type: :digital_output,
+      endpoints:
+        Enum.map(@channels, &%Endpoint{signal: &1, name: &1, direction: :output, type: :boolean}),
+      commands: []
+    }
+  end
 
   @impl true
   def project_state(_decoded_inputs, prev_state, driver_state, _config)